bugfix: formatter

+++ frontend/src/assets/thumbnailFormatter.py
@@ -4,7 +4,8 @@ import re
 import os
 
 for path in glob.glob("thumbnails/*"):
-    img = Image.open(path).convert('RGB')
-    number = max(re.findall(r"\d+", path))
-    img.save(f"thumbnails/{number}.png", "PNG")
-    os.remove(path)
+    if not "png" in path:
+        img = Image.open(path).convert('RGB')
+        number = max(re.findall(r"\d+", path))
+        img.save(f"thumbnails/{number}.png", "PNG")
+        os.remove(path)
